- Code cleaning - Add TAP Testing framework
