+++ backend/package.json
@@ -17,6 +17,7 @@
     "@typescript-eslint/eslint-plugin": "^6.18.0",
     "@typescript-eslint/parser": "^6.18.0",
     "eslint": "^8.56.0",
+    "tap": "^18.6.1",
     "tsc-watch": "^6.0.4",
     "typescript": "^5.3.3"
   },
