+++ backend/src/entry.ts
@@ -1,129 +1,7 @@
 import process from "process";
-
-import fastifyCors from "@fastify/cors";
-import fastifyMultipart from "@fastify/multipart";
-import fastifyStatic from "@fastify/static";
-import Fastify from "fastify";
-
-import { default as BookmarkController } from "./Controllers/bookmarks";
-import { default as DirController } from "./Controllers/dir";
-import { default as FsController } from "./Controllers/fs";
-import { default as UploadController } from "./Controllers/upload";
-import { default as UserController } from "./Controllers/user";
-import { default as SettingsController } from "./Controllers/settings";
-
-import { error, info } from "npmlog";
-import path from "path";
-import authMiddleware from "./authMiddleware";
-import { customErrorHandler } from "./errorHandler";
-import { BANNED_IP_TIMEOUT, SESSIONS } from "./global";
-
-/* --------------------------------- CONSTS --------------------------------- */
-
-export const PATH_PREFIX = process.env?.["FM_PATH_PREFIX"] ?? "/";
-export const SESSION_TIMEOUT = 1000 * 3600 * 12;
-
-/* ------------------------------ INIT FASTIFY ------------------------------ */
-
-const fastify = Fastify({
-  logger: false
-});
-
-fastify.register(fastifyMultipart, {
-  limits: {
-    fileSize: 200 * 1000 * 1000
-  },
-});
-
-if (process.env?.["EXP_MAX_CHUNK_SIZE"] && +process.env["EXP_MAX_CHUNK_SIZE"] > 200) {
-  error("Config", "EXP_MAX_CHUNK_SIZE cannot be more than 200");
-  process.exit(1);
-}
-
-fastify.register(fastifyCors, {
-  origin: "*"
-});
-
-fastify.register(fastifyStatic, {
-  root: process.env["NODE_ENV"] === "production" ? "/front/dist" : path.join(__dirname, "../../front/dist"),
-  // prefix: PATH_PREFIX,
-  prefix: PATH_PREFIX === "/" ? "/" : PATH_PREFIX + "/",
-});
-
-
-fastify.register(authMiddleware);
-fastify.setErrorHandler(customErrorHandler);
-
-/* ----------------------- HOOK FOR REVERSE-PROXY PATH ---------------------- */
-
-// TODO: Rework
-fastify.addHook("onSend", async function (req, rep, payload: any) {
-  const url = req.url.replace(PATH_PREFIX, "");
-
-  if (
-    PATH_PREFIX === "/" ||
-    url.startsWith("api") ||
-    (url !== "" && !url.endsWith(".css")
-      && !url.endsWith(".html")
-      && !url.endsWith(".js")) ||
-    typeof payload.on !== "function"
-  ) {
-    return payload;
-  }
-
-  return new Promise((resolve, reject) => {
-    const chunks = [];
-
-    payload.on("error", (err) => {
-      reject(err);
-    });
-
-    payload.on("data", (data) => {
-      // console.log(data);
-      chunks.push(data);
-    });
-
-    payload.on("end", () => {
-      let str = chunks.join("");
-
-      const prefix = PATH_PREFIX.startsWith("/") ? PATH_PREFIX.slice(1) : PATH_PREFIX;
-
-      if (url.endsWith("js") || url.endsWith("css")) {
-        str = str.replaceAll(
-          "assets/",
-          prefix + "/assets/"
-        );
-        str = str.replaceAll(
-          "/resources/",
-          PATH_PREFIX + "/resources/"
-        );
-      } else {
-        str = str.replaceAll(
-          "/assets/",
-          PATH_PREFIX + "/assets/"
-        );
-      }
-      //@ts-expect-error ---
-      resolve(str);
-    });
-  });
-});
-
-/* --------------------------------- ROUTES --------------------------------- */
-
-fastify.get(PATH_PREFIX, function (_, res) {
-  return res.sendFile("index.html");
-});
-
-
-fastify.register(DirController, { prefix: "/api/dir" });
-fastify.register(FsController, { prefix: "/api/fs" });
-fastify.register(UploadController, { prefix: "/api/upload" });
-fastify.register(UserController, { prefix: "/api/users" });
-fastify.register(BookmarkController, { prefix: "/api/bookmarks" });
-fastify.register(SettingsController, { prefix: "/api/settings" });
-
-/* ------------------------------- SERVER MAIN ------------------------------ */
+import { info } from "npmlog";
+import { fastifyServer } from "./fastify";
+import { BANNED_IP_TIMEOUT, PATH_PREFIX, SESSIONS, SESSION_TIMEOUT } from "./global";
 
 const SIGNALS = {
   "SIGHUP": 1,
@@ -138,7 +16,7 @@ const SIGNALS = {
       process.on(signal, () => {
         console.log(`process received a ${signal} signal`);
         // shutdown(signal, signals[signal]);
-        fastify.close(() => {
+        fastifyServer.close(() => {
           console.log(`server stopped by ${signal} with value ${SIGNALS[signal]}`);
           process.exit(128 + SIGNALS[signal]);
         });
@@ -163,9 +41,10 @@ const SIGNALS = {
       }
     }, 60000 * 5);
 
-    await fastify.listen({ port: 3000, host: "0.0.0.0" });
+    await fastifyServer.listen({ port: 3000, host: "0.0.0.0" });
   } catch (err) {
-    fastify.log.error(err);
+    fastifyServer.log.error(err);
     process.exit(1);
   }
 })();
+
